+++ lib/Models/Workflows/SelectableDimensionWorkflow.ts
@@ -4,7 +4,10 @@ import { BaseModel } from "../Definition/Model";
 import {
   SelectableDimensionGroup,
   SelectableDimension
-} from "./SelectableDimensions";
+} from "../SelectableDimensions/SelectableDimensions";
+import ViewState from "../../ReactViewModels/ViewState";
+import Terria from "../Terria";
+import { runInAction } from "mobx";
 
 /**
  * Model for SelectableDimensionWorkflow. Basically just includes a bunch of selectableDimensions, name, icon and a catalog member.
@@ -16,13 +19,35 @@ export default interface SelectableDimensionWorkflow {
   /** Human readable name - used as title */
   name: string;
   icon: IconProps["glyph"];
+
+  /** Item to which this workflow belongs **/
   item: BaseModel;
+
   onClose?: () => void;
   /** Footer button */
   footer?: { onClick: () => void; buttonText: string };
   menu?: PanelMenuProps;
   /** This allows up to two levels of SelectableDimensionGroup */
   selectableDimensions: SelectableDimensionWorkflowGroup[];
+}
+
+/**
+ * Runs a selectable dimension workflow which is a workflow for a workbench item.
+ *
+ * @param viewStateOrTerria - The {@link ViewState} or {@link Terria} instance
+ * @param workflow - A {@link SelectableDimensionWorkflow} instance
+ */
+export function runWorkflow(
+  viewStateOrTerria: ViewState | Terria,
+  workflow: SelectableDimensionWorkflow
+) {
+  runInAction(() => {
+    const terria =
+      viewStateOrTerria instanceof Terria
+        ? viewStateOrTerria
+        : viewStateOrTerria.terria;
+    terria.selectableDimensionWorkflow = workflow;
+  });
 }
 
 /** This is essentially the same as `SelectableDimensionGroup`, but allows two levels of nested `SelectableDimensionGroup`, instead of one */